UWM-CMS: Custom module to set X-Robots-Tag header (#700)
* UWM-CMS: Custom module to set X-Robots-Tag header

If a request to stevie or uwmed does not come from the host of "www.uwmedicine.org" then set the X-Robots-Tag header to "noindex,nofollow". The goal is to prevent stevie.cms and cms. URLs from being indexed in Google.

This commit also includes an update to the default metatag settings to set the base URL to https://www.uwmedicine.org which will update the canonical meta tags and headers.

* UWM-CMS: Fix whitespace linting error